+++ psynet/serialize.py
@@ -129,8 +129,10 @@ def _call_trial_maker_method(
 
 
 def _ensure_sql_primary_key(instance, context: str) -> None:
-    handler = SQLHandler()
-    primary_keys = handler.get_primary_keys(instance)
+    primary_key_cols = [
+        column.name for column in instance.__class__.__table__.primary_key.columns
+    ]
+    primary_keys = {key: getattr(instance, key) for key in primary_key_cols}
     if any(key is None for key in primary_keys.values()):
         raise CallbackSerializationError(
             _format_callback_error(